Bugfix: Ok the header parser did not even remotely what it was supposed to do. Now it does.
Additionally: Fix the parsing loop to consume as much data as possible
in one run. This was made obvious when using POST requests with Apple's
HTTP client implementation. It seems to send the Intro and Headers in
one TCP packet and the body in another, even if the MTU was not reached
so the parser would sometimes hang.